Taking code from pair project, modifying it for my needs
I’m taking files from the pair project (the blog) and modifying the
code so that it suits this project. To truly write it from scratch is
somewhat daunting .. at least this way I have a template/some
structure. projects.js will contain the array of objects that I will be
using - the objects are each portfolio piece.